Established portfolio structure & content
Utilised portfolYOU's custom theme from: https://github.com/yousinix/portfolYOU as a starting point for a Jekyll themed portfolio that works with GitHub Pages. Adjusted large amounts of the theme and styling to be more visually cohesive and pleasing. Added the beginnings of some of the content - added CART/DRFT as well as the OG shader study of Slime Rancher from my Squarespace, alongside custom content for landing and about pages.